+++ controllers/keyController.js
@@ -3,8 +3,16 @@ class KeyController {
         this.keyService = keyService;
     }
 
-    generateKey() {
+    createHash() {
         return this.keyService.generateKey();
+    }
+
+    createManyHashes(count) {
+        return this.keyService.generateMultipleKeys(count);
+    }
+
+    createCustomHash(customString) {
+        return this.keyService.generateCustomKey(customString);
     }
 }